Fix C4242 warning in serial_port_base::character_size::store.
Add static_cast<BYTE> to avoid MSVC warning C4242 when converting
from unsigned int to BYTE in character_size::store(). This warning
causes issues for projects that treat warnings as errors.

The cast is safe since character_size values are always in the valid
BYTE range (typically 5, 6, 7, or 8).

+++ include/asio/impl/serial_port_base.ipp
@@ -508,7 +508,7 @@ ASIO_SYNC_OP_VOID serial_port_base::character_size::store(
     ASIO_OPTION_STORAGE& storage, asio::error_code& ec) const
 {
 #if defined(ASIO_WINDOWS) || defined(__CYGWIN__)
-  storage.ByteSize = value_;
+  storage.ByteSize = static_cast<BYTE>(value_);
 #else
   storage.c_cflag &= ~CSIZE;
   switch (value_)
